Prettify error output for ARM deployment
